+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   <summary>
     Changes that have landed in master but are not yet released.
     Click to see more.
+    
+    - Add readOnly prop to CustomFieldInputCurrency and CustomFieldInputNumber
   </summary>
 </details>